requirements.txt
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.3
grpcio==1.78.1
grpcio-tools==1.78.1
idna==3.11
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
platformdirs==4.9.2
protobuf==6.33.5
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
setuptools==82.0.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.6
zeep==4.3.2
